move skip-eviction annotation into the sdk
