+++ sdk/node/eviction.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package types
+package node
 
 const (
 	SkipEvictionAnnotationKey = "kubermatic.io/skip-eviction"
